% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{add_metrics_to_dataframe}
\alias{add_metrics_to_dataframe}
\title{Fill one row of data.frame df with this site's metrics}
\usage{
add_metrics_to_dataframe(df, ts, tel_test, df_idx, AC_rating,
  forecast_type, runtime)
}
\arguments{
\item{df}{A data frame to fill with metrics for all the sites}

\item{ts}{A ts_forecast object}

\item{tel_test}{Vector of telemetry for the test period, corresponding to the length of ts}

\item{df_idx}{Index of data frame row (likely index of site)}

\item{AC_rating}{Site's AC rating}

\item{forecast_type}{Name of forecast method, e.g., "raw"}

\item{runtime}{A value in seconds}
}
\value{
data.frame df with a row added for this site
}
\description{
CRPS, interval width, and MAE are normalized by the site's AC power rating.
}
