% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_maxar_parameters}
\alias{check_maxar_parameters}
\title{Subfunction to error-check temporal parameters for Maxar load-in}
\usage{
check_maxar_parameters(nc, metadata, site)
}
\arguments{
\item{nc}{An open NetCDF object}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}

\item{site}{Site index}
}
\description{
Subfunction to error-check temporal parameters for Maxar load-in
}
