% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{export_quantiles_to_h5}
\alias{export_quantiles_to_h5}
\title{Export a [issue time x step x quantile] array of quantiles to HDF5 format}
\usage{
export_quantiles_to_h5(forecast_runs, fname)
}
\arguments{
\item{forecast_runs}{A list of ts_forecast objects}

\item{fname}{Path to output file}
}
\description{
Quantile forecasts are extracted from the list of ts_forecast objects and
exported in array format to an HDF5 file. Can equivalently export a file
with a NetCDF extension; ncdf4 package is used for NetCDF/HDF5 handling.
}
