% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_by_site.R
\name{forecast_by_site}
\alias{forecast_by_site}
\title{Conduct site-level forecasting with the desired method}
\usage{
forecast_by_site(t_idx_series, ens_data, all_site_tel, sun_up, site,
  AC_rating, metadata, lm_formula, runtime_data_dir)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
all_site_tel time indices}

\item{ens_data}{[site x member x time] array of all ensemble data (historical
+ test)}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{sun_up}{A [site x time] matrix of booleans}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}

\item{runtime_data_dir}{Directory to save .Rdata results}
}
\description{
For the given forecasting method, this function conducts the appropriate
training method and forecasts for the test period. Exports .Rdata file of
results to the "Runtime data" directory
}
