% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bma_ts.R
\name{get_bma_ts}
\alias{get_bma_ts}
\title{Get time-series forecast using BMA post-processing of NWP ensembles}
\usage{
get_bma_ts(t_idx_series, ens_test, ensemble, telemetry, sun_up, site,
  AC_rating, metadata, lm_formula)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
telemetry time indices}

\item{ens_test}{[time x member] matrix of ensemble data for test period}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}
}
\value{
A ts_forecast object
}
\description{
This method trains BMA models for each time-point using a sliding,
time-of-day, or constant training method, then forecasts using the current
NWP ensemble.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_chpeen_ts}}, \code{\link{get_clim_ts}},
  \code{\link{get_emos_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
