% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clim_ts.R
\name{get_clim_ts}
\alias{get_clim_ts}
\title{Get time-series forecast using climatology}
\usage{
get_clim_ts(tel_test, sun_up, sites, site_idx, site_max_p, metadata)
}
\arguments{
\item{tel_test}{A vector of the test period telemetry for this site}

\item{sun_up}{A [site x time] matrix of booleans}

\item{metadata}{A data.frame of forecast parameters}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}
}
\value{
A ts_forecast object
}
\description{
This method generates a static forecast for all times when the sun is up,
*based on the climatology of the test sample*, rather than a historical
period.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_chpeen_ts}},
  \code{\link{get_emos_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
