% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emos_ts.R
\name{get_emos_ts}
\alias{get_emos_ts}
\title{Get time-series forecast using EMOS post-processing of NWP ensembles}
\usage{
get_emos_ts(t_idx_series, ens_test, ens_data, all_site_tel, sun_up, sites,
  site_idx, site_max_p, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
all_site_tel time indices}

\item{ens_test}{[time x member] matrix of ensemble data for test period}

\item{ens_data}{[site x member x time] array of all ensemble data (historical
+ test)}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{sun_up}{A [site x time] matrix of booleans}

\item{metadata}{A data.frame of forecast parameters}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}
}
\value{
A ts_forecast object
}
\description{
This method trains EMOS models for each time-point using a sliding or
time-of-day training method, then forecasts using the current
NWP ensemble.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_chpeen_ts}},
  \code{\link{get_clim_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
