% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_forecast_data}
\alias{get_forecast_data}
\title{Load ensemble forecast data}
\usage{
get_forecast_data(fname, members, site, metadata, date_start, ...)
}
\arguments{
\item{fname}{file name}

\item{members}{A vector of member indices}

\item{site}{Site index}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}

\item{date_start}{Timestamp}

\item{...}{Additional parameters to load-in subfunctions}
}
\value{
A list of data=[issue x step x member] matrix and issuetime=a
  vector of POSIXct timestamps
}
\description{
Loads in ensemble forecast data into the form: [issue x step x member]
Also generates a list of validtime timestamps
}
\details{
If input file is in Maxar form, assumes a NETCDF file of the dimensions: [Day
x Hour x Site x Lead time x member] Maxar data can be loaded to either match
the form above or in a "rolling" format over the course of the year in the
form [1 x all steps x member] to make annual average metrics easier

If input file in in ECMWF format (via reV), assumes an h5 file of the
dimensions: [issue x step x member]
}
