% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_hourly_pvw_ensemble}
\alias{get_hourly_pvw_ensemble}
\title{Subfunction to load in hourly PVW forecasts}
\usage{
get_hourly_pvw_ensemble(nc, site, date_start, metadata, vname = "power",
  date_data_start = lubridate::ymd_h("20170101_12"))
}
\arguments{
\item{nc}{An open NetCDF object}

\item{site}{Site index}

\item{date_start}{Timestamp}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}

\item{vname}{NetCDF variable name}

\item{date_data_start}{A lubridate: Date of first day in file, already in UTC}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}
}
\description{
Loads in ensemble forecast data into the form: [issue x step x member]
Assumes a NETCDF file of the dimensions: [site x issue x step x member]
Loads in all members by default
}
