% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_maxar_data_by_horizon}
\alias{get_maxar_data_by_horizon}
\title{Subfunction to get single data point by issue and horizon}
\usage{
get_maxar_data_by_horizon(h, issue, member, data_rectangle, date_start,
  metadata)
}
\arguments{
\item{h}{Horizon index}

\item{issue}{A lubridate: issue time and hour}

\item{member}{Member index}

\item{data_rectangle}{Array of data from the NetCDF file}

\item{date_start}{A lubridate: Start date of data to load}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}
}
\description{
Presumes 1-hour resolution of indices in data_rectangle
}
\keyword{internal}
