% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_maxar_data_by_issue}
\alias{get_maxar_data_by_issue}
\title{Subfunction to get diagonal data by issue time}
\usage{
get_maxar_data_by_issue(member, data_rectangle, ensemble_issue_times,
  metadata)
}
\arguments{
\item{member}{Member index}

\item{data_rectangle}{Array of data from the NetCDF file}

\item{ensemble_issue_times}{A sequence of lubridate issue times}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}
}
\description{
Subfunction to get diagonal data by issue time
}
\keyword{internal}
