% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_maxar_ensemble}
\alias{get_maxar_ensemble}
\title{Subfunction to load in ECMWF data}
\usage{
get_maxar_ensemble(nc, members, site, metadata, ensemble_issue_times,
  vname = "powernew", truncate = T,
  date_data_start = lubridate::ymd(20160101), AC_rating = NULL)
}
\arguments{
\item{nc}{An open NetCDF object}

\item{members}{A vector of member indices}

\item{site}{Site index}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}

\item{ensemble_issue_times}{A sequence of lubridate issue times}

\item{vname}{NetCDF variable name}

\item{truncate}{Boolean: Whether or not to truncate the forecasts at the site
maximum power}

\item{date_data_start}{A lubridate: Date of first day in file}

\item{AC_rating}{AC power rating}
}
\description{
Loads in ensemble forecast data into the form: [issue x step x member]
If input file is in Maxar form, assumes a NETCDF file of the dimensions: [Day
x Hour x Site x Lead time x member] Maxar data can be loaded to either match
the form above or in a "rolling" format over the course of the year in the
form [1 x all steps x member] to make annual average metrics easier
}
