% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_maxar_telemetry}
\alias{get_maxar_telemetry}
\title{Load data from a NETCDF file of telemetry}
\usage{
get_maxar_telemetry(nc, site, metadata, date_start,
  date_data_start = lubridate::ymd(20160101), vname = "hsl_power")
}
\arguments{
\item{nc}{Open NetCDF file}

\item{site}{Site index}

\item{metadata}{Metadata list including date end, temporal parameters,
time-steps per day, rolling or not, etc.}

\item{date_start}{A lubridate: Start date of data to load}

\item{date_data_start}{A lubridate: Date of first day in file}

\item{vname}{NetCDF variable name}
}
\value{
A vector of telemetry
}
\description{
Assumed file dimensions: Day x Hour x Site
Time-point selection is a consecutive sequence
}
