% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_metrics_for_single_run}
\alias{get_metrics_for_single_run}
\title{Get a vector of the average metrics over a single forecast run}
\usage{
get_metrics_for_single_run(i, forecast_runs, issue_times, metadata,
  telemetry)
}
\arguments{
\item{i}{Index of forecast run in the list of forecast_runs and issue_times}

\item{forecast_runs}{A list of ts_forecast object}

\item{issue_times}{A list of timestamps, same length as forecast_runs}

\item{metadata}{A data.frame of forecast parameters}

\item{telemetry}{Vector of telemetry}
}
\value{
A vector of average metrics
}
\description{
Get a vector of the average metrics over a single forecast run
}
