% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_ndays}
\alias{get_ndays}
\title{Calculate number of days in the sequence}
\usage{
get_ndays(date_start, date_end)
}
\arguments{
\item{date_start}{A lubridate: Start date of data to load}

\item{date_end}{A lubridate: End date of data to load}
}
\value{
Number of days in requested data sequence
}
\description{
Calculate number of days in the sequence
}
