% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_quantile_array}
\alias{get_quantile_array}
\title{Extract an [issue time x step x quantile] array of quantile forecasts}
\usage{
get_quantile_array(forecast_runs)
}
\arguments{
\item{forecast_runs}{A list of ts_forecast objects}
}
\value{
a [issue time x step x quantile] array
}
\description{
Quantile forecasts are extracted from the list of ts_forecast objects (each
an individual forecast run), at the resolution within the prob_forecast
objects (defaults to 0.1%)
}
