% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_ts.R
\name{get_raw_ts}
\alias{get_raw_ts}
\title{Get time-series forecast of raw NWP ensemble}
\usage{
get_raw_ts(issue, ens_test, site, AC_rating, metadata)
}
\arguments{
\item{issue}{A time stamp}

\item{ens_test}{[time x member] matrix of ensemble data for test period}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}
}
\value{
A ts_forecast object
}
\description{
This method does a basic empirical CDF based on the raw NWP ensemble.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_chpeen_ts}},
  \code{\link{get_clim_ts}}, \code{\link{get_emos_ts}},
  \code{\link{get_peen_ts}}
}
\concept{ts_training_forecast}
