% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_start_day}
\alias{get_start_day}
\title{Calculate start day's index since the beginning of data availability}
\usage{
get_start_day(date_data_start, date_start)
}
\arguments{
\item{date_data_start}{A lubridate: Date of first day in file}

\item{date_start}{A lubridate: Start date of data to load}
}
\value{
Index number of first requested day
}
\description{
Calculate start day's index since the beginning of data availability
}
