% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_training_ensemble_from_validtimes}
\alias{get_training_ensemble_from_validtimes}
\title{Subset ensemble data to get BMA/EMOS training data}
\usage{
get_training_ensemble_from_validtimes(time_idx_train, ensemble, metadata)
}
\arguments{
\item{time_idx_train}{Vector of time-point indices, relative to telemetry's valid times}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{metadata}{A data.frame of forecast parameters}
}
\description{
This function will need to be expanded for full functionality of non-rolling forecasts
}
