% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_training_subsets}
\alias{get_training_subsets}
\title{Get subsets of ensemble and telemetry data for BMA/EMOS training}
\usage{
get_training_subsets(time_idx_forecast, issue, step, metadata, ensemble,
  telemetry)
}
\arguments{
\item{time_idx_forecast}{Index of forecast time, relative to telemetry's
valid times}

\item{step}{Step (index) in this forecast run}

\item{metadata}{A data.frame of forecast parameters}

\item{ensemble}{A list of data=[issue x step x member] array of all ensemble
data (historical + test) and issuetime=vector of POSIXct time stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}
}
\description{
Implements different strategies based on on sliding/time-of-day training and
rolling/non-rolling formats. For non-rolling forecasts, sliding training
generates only a single model at the issue time using the sequence of data up
until one step before issue. Time-of-day training generates unique models for
each step in the run, matching the observations from previous days to the
forecast issued at the same total lookahead time (lead time + partial
horizon).
}
