% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{issue_2_valid_index}
\alias{issue_2_valid_index}
\title{Translate ensemble issue time and step to telemetry valid time index}
\usage{
issue_2_valid_index(issue, step, metadata, telemetry)
}
\arguments{
\item{issue}{A POSIXct timestamp of issue time}

\item{step}{Index of forecast step in this run}

\item{metadata}{A data.frame of forecast parameters}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}
}
\value{
an index of the telemetry$data vector
}
\description{
Translate ensemble issue time and step to telemetry valid time index
}
