% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bma_ts.R
\name{train_bma_subfunc}
\alias{train_bma_subfunc}
\title{Subfunction to get BMA model for individual time-point using sliding or
time-of-day training}
\usage{
train_bma_subfunc(time_idx_forecast, ensemble, telemetry, sun_up, site,
  AC_rating, metadata, lm_formula)
}
\arguments{
\item{time_idx_forecast}{Time index}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}
}
\description{
Subfunction to get BMA model for individual time-point using sliding or
time-of-day training
}
