% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chpeen_ts.R
\name{train_ch_peen}
\alias{train_ch_peen}
\title{Subfunction to get input data for CH-PeEn forecast}
\usage{
train_ch_peen(t_idx_series, telemetry, sun_up, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
telemetry time indices}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{metadata}{A data.frame of forecast parameters}
}
\description{
Get matrix of static, complete-history PeEn data
}
