% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emos_ts.R
\name{train_emos}
\alias{train_emos}
\title{Subfunction to get EMOS models using sliding or time-of-day training}
\usage{
train_emos(t_idx_series, ens_data, all_site_tel, sun_up, sites, site_idx,
  site_max_p, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
all_site_tel time indices}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{sun_up}{A [site x time] matrix of booleans}

\item{metadata}{A data.frame of forecast parameters}

\item{ens_test}{[time x member] matrix of ensemble data for test period}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}
}
\description{
Subfunction to get EMOS models using sliding or time-of-day training
}
