% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emos_ts.R
\name{train_emos_subfunc}
\alias{train_emos_subfunc}
\title{Subfunction to get EMOS model for individual time-point using sliding or
time-of-day training}
\usage{
train_emos_subfunc(step, t_idx_series, issue, ensemble, telemetry, sun_up,
  site, AC_rating, metadata)
}
\arguments{
\item{step}{Step (index) in this forecast run}

\item{t_idx_series}{Series of time indices in forecast run, relative to the
telemetry time indices}

\item{issue}{Time stamp of issue time}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}
}
\description{
Subfunction to get EMOS model for individual time-point using sliding or
time-of-day training
}
