% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peen_ts.R
\name{train_peen_data}
\alias{train_peen_data}
\title{Get matrix of evolving PeEn data}
\usage{
train_peen_data(t_idx_series, all_site_tel, sun_up, site_idx, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
all_site_tel time indices}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{sun_up}{A [site x time] matrix of booleans}

\item{site_idx}{Index of which site is currently being forecasted, relative
to the order in all_site_tel}

\item{metadata}{A data.frame of forecast parameters}
}
\description{
Get matrix of evolving PeEn data
}
