% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{valid_2_issue_index}
\alias{valid_2_issue_index}
\title{Translate telemetry valid time to ensemble issue/step indices}
\usage{
valid_2_issue_index(valid, metadata, ensemble, issue = NULL)
}
\arguments{
\item{valid}{A POSIXct timestamp of valid time}

\item{metadata}{A data.frame of forecast parameters}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{issue}{(optional) A POSIXct timestamp of issue time, defaults to most
recent forecast}
}
\value{
c(issue, step) indices of the ensemble$data vector
}
\description{
Translate telemetry valid time to ensemble issue/step indices
}
