% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{export_metrics_to_csv}
\alias{export_metrics_to_csv}
\title{Export forecasting metrics to .csv}
\usage{
export_metrics_to_csv(out_dir, runtime_data_dir, sites, site_max_p,
  metadata, export_plots = F)
}
\arguments{
\item{out_dir}{Main output directory to save graphs}

\item{runtime_data_dir}{Directory where .Rdata results are saved}

\item{sites}{Vector of sites that are being forecast in this simulations}

\item{site_max_p}{A vector of maximum AC power rating for all sites in this
dataset, regardless of which subset in "sites" is being simulated}

\item{metadata}{A data.frame of forecast parameters}

\item{export_plots}{Boolean flag (default=False)}

\item{site_idx}{Index of which site is currently being forecasted, relative
to the order in all_site_tel}
}
\description{
Cycles through site-level forecasts to assess metrics, including CRPS
(weighted and unweighted) and data quality metrics. If "export_plots" is
selected, this will generate reliability, sharpness, quantile score, and PIT
histogram plots *for each site*.
}
