% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_by_issue_time.R
\name{forecast_by_issue_time}
\alias{forecast_by_issue_time}
\title{Conduct site-level forecasting with the desired method}
\usage{
forecast_by_issue_time(issue, ensemble, telemetry, sun_up, site, max_power,
  metadata, lm_formula)
}
\arguments{
\item{issue}{A lubridate time stamp of the issue time}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{An [issue x step] matrix of booleans}

\item{site}{String, site name}

\item{max_power}{Maximum load or site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}
}
\description{
For the given forecasting method, this function conducts the appropriate
training method and forecasts for the test period. Exports .Rdata file of
results to the "Runtime data" directory
}
