% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binned_ts.R
\name{get_binned_ts}
\alias{get_binned_ts}
\title{Get time-series forecast of binned probability forecast}
\usage{
get_binned_ts(issue, ens_test, site, max_power, metadata)
}
\arguments{
\item{issue}{A time stamp}

\item{ens_test}{[time x member] matrix of ensemble data}

\item{site}{String, site name}

\item{max_power}{Site's AC power rating or maximum load}

\item{metadata}{A data.frame of forecast parameters}
}
\value{
A ts_forecast object
}
\description{
This method does a basic linear interpolation of the raw NWP ensemble.
}
\seealso{
Other ts_training_forecast: \code{\link{get_bma_ts}},
  \code{\link{get_chpeen_ts}}, \code{\link{get_clim_ts}},
  \code{\link{get_emos_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
