% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chpeen_ts.R
\name{get_chpeen_ts}
\alias{get_chpeen_ts}
\title{Get time-series forecast using CH-PeEn method}
\usage{
get_chpeen_ts(issue, t_idx_series, telemetry, sun_up, site, AC_rating,
  metadata)
}
\arguments{
\item{issue}{A time stamp}

\item{t_idx_series}{Series of time indices to forecast, relative to the
telemetry time indices}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}
}
\value{
A ts_forecast object
}
\description{
This method generates a static 24-hour or similar forecast using all data at
the same time of day in the training period only. This is a static forecast
that does not evolve as more observations are realized in the evaluation
period.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_clim_ts}},
  \code{\link{get_emos_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
