% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clim_ts.R
\name{get_clim_ts}
\alias{get_clim_ts}
\title{Get time-series forecast using climatology}
\usage{
get_clim_ts(issue, telemetry, sun_up, site, AC_rating, metadata)
}
\arguments{
\item{issue}{A time stamp}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}
}
\value{
A ts_forecast object
}
\description{
This method generates a static forecast for all times when the sun is up.
The training telemetry could either be the test sample or a historical
period.
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_chpeen_ts}},
  \code{\link{get_emos_ts}}, \code{\link{get_peen_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
