% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peen_ts.R
\name{get_peen_ts}
\alias{get_peen_ts}
\title{Get time-series forecast using persistence ensemble (PeEn) method}
\usage{
get_peen_ts(t_idx_series, telemetry, sun_up, site, AC_rating, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
telemetry time indices}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}
}
\value{
A ts_forecast object
}
\description{
This method enforces a set of training data of length training_window; if
there is missing data, it continues back through the historical record until
it has a full set (unless it runs out of data)
}
\seealso{
Other ts_training_forecast: \code{\link{get_binned_ts}},
  \code{\link{get_bma_ts}}, \code{\link{get_chpeen_ts}},
  \code{\link{get_clim_ts}}, \code{\link{get_emos_ts}},
  \code{\link{get_raw_ts}}
}
\concept{ts_training_forecast}
