% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bma_ts.R
\name{train_bma}
\alias{train_bma}
\title{Subfunction to get BMA models using sliding or time-of-day training}
\usage{
train_bma(t_idx_series, ensemble, telemetry, sun_up, site, AC_rating,
  metadata, lm_formula)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
telemetry time indices}

\item{ensemble}{A list of data=[issue x step x member] array of all ensemble
data (historical + test) and issuetime=vector of POSIXct time stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}
}
\description{
There are four possible training methods, with combinations of
sliding/time-of-day and rolling/non-rolling forecasts. Sliding/non-rolling
forecasts will only train a single model at the issue time, then apply that
model over the forecast run. The other three combinations will generate
unique models at each valid time.
}
