% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bma_ts.R
\name{train_bma_subfunc}
\alias{train_bma_subfunc}
\title{Subfunction to get BMA model for individual time-point using sliding or
time-of-day training}
\usage{
train_bma_subfunc(time_idx_forecast, ens_data, all_site_tel, sun_up, site,
  AC_rating, metadata, lm_formula)
}
\arguments{
\item{time_idx_forecast}{Time index}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{sun_up}{A [site x time] matrix of booleans}

\item{site}{String, site name}

\item{AC_rating}{Site's AC power rating}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}

\item{ens_test}{[time x member] matrix of ensemble data for test period}
}
\description{
Subfunction to get BMA model for individual time-point using sliding or
time-of-day training
}
