% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bma_ts.R
\name{train_bma_subfunc}
\alias{train_bma_subfunc}
\title{Subfunction to get BMA model for individual time-point using sliding or
time-of-day training}
\usage{
train_bma_subfunc(step, t_idx_series, issue, ensemble, telemetry, sun_up,
  site, max_power, metadata, lm_formula)
}
\arguments{
\item{step}{Step (index) in this forecast run}

\item{t_idx_series}{Series of time indices in forecast run, relative to the
telemetry time indices}

\item{issue}{Time stamp of issue time}

\item{ensemble}{A list of data=[issue x step x member] array of all
ensemble data (historical + test) and issuetime=vector of POSIXct time
stamps}

\item{telemetry}{A list of data=vector of telemetry and validtime=vector of
POSIXct times}

\item{sun_up}{A vector of booleans, indexed by telemetry valid times}

\item{site}{String, site name}

\item{max_power}{Site's AC power rating or maximum load}

\item{metadata}{A data.frame of forecast parameters}

\item{lm_formula}{Formula for BMA linear regression}
}
\description{
Subfunction to get BMA model for individual time-point using sliding or
time-of-day training
}
