% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chpeen_ts.R
\name{train_ch_peen}
\alias{train_ch_peen}
\title{Subfunction to get input data for CH-PeEn forecast}
\usage{
train_ch_peen(t_idx_series, all_site_tel, sun_up, site_idx, metadata)
}
\arguments{
\item{t_idx_series}{Series of time indices to forecast, relative to the
all_site_tel time indices}

\item{all_site_tel}{A [site x time] matrix of telemetry}

\item{metadata}{A data.frame of forecast parameters}

\item{site}{String, site name}
}
\description{
Get matrix of static, complete-history PeEn data
}
